python_pipeline/requirements.txt
opencv-python
scikit-learn
numpy
pandas
pillow
imagehash
face-recognition
exifread
sqlite3-binary
